Updates:  - refactor both models and blocks into their own modules to reduce the size of large python modules  - removed the redundant link in cards template  - additional gap at top of footer for non-mobile view  - added "tertiary" support extending bootstrap
