+++ cmspage/__init__.py
@@ -1,0 +1,22 @@
+DEFAULT_RICHTEXTBLOCK_FEATURES = [
+    "h2",
+    "h3",
+    "h4",
+    "h5",
+    "h6",
+    "bold",
+    "italic",
+    "ol",
+    "ul",
+    "hr",
+    "link",
+    "document-link",
+    "image",
+    "embed",
+    "code",
+    "blockquote",
+    "superscript",
+    "subscript",
+    "strikethrough",
+    "usefont",
+]
