+++ cmspage/models/choice_icon.py
