+++ cmspage/scss/partials/_variables.scss
@@ -3,3 +3,5 @@ $breakpoint-md: 768px;
 $breakpoint-lg: 992px;
 $breakpoint-xl: 1200px;
 $breakpoint-xxl: 1400px;
+
+$tertiary: #a7c6d9d4;
