+++ cmspage/templates/cmspage/cms_footer_page.html
@@ -20,7 +20,7 @@
   <body>
 {% endif %}
 
-<footer class="container footer footer-content">
+<footer class="container footer footer-content mt-md-2">
     {% for block in page.footer %}
       {% include_block block %}
     {% empty %}
